requirements.txt
sqlmodel>=0.0.8
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
asyncpg>=0.24.0 